# Add Indexes to candles and orders to optimize queries