--- a/indexer/packages/postgres/src/db/migrations/migration_files/20240208153050_add_candles_ticket_resolution_index.ts
+++ b/indexer/packages/postgres/src/db/migrations/migration_files/20240208153050_add_candles_ticket_resolution_index.ts
@@ -1,0 +1,18 @@
+import * as Knex from 'knex';
+
+export async function up(knex: Knex): Promise<void> {
+  // eslint-disable-next-line @typescript-eslint/quotes
+  await knex.raw(`
+    CREATE INDEX CONCURRENTLY IF NOT EXISTS candles_ticker_resolution_index ON candles("ticker", "resolution");
+  `);
+}
+
+export async function down(knex: Knex): Promise<void> {
+  await knex.raw(`
+    DROP INDEX CONCURRENTLY IF EXISTS "candles_ticker_resolution_index";
+  `);
+}
+
+export const config = {
+  transaction: false,
+};