--- a/indexer/packages/postgres/src/db/migrations/migration_files/20240208161948_add_orders_clobpairid_subaccountid_index.ts
+++ b/indexer/packages/postgres/src/db/migrations/migration_files/20240208161948_add_orders_clobpairid_subaccountid_index.ts
@@ -3,13 +3,13 @@ import * as Knex from 'knex';
 export async function up(knex: Knex): Promise<void> {
   // eslint-disable-next-line @typescript-eslint/quotes
   await knex.raw(`
-    CREATE INDEX CONCURRENTLY IF NOT EXISTS orders_clobpairid_subaccountid_index ON orders("clobPairId", "subaccountId");
+    CREATE INDEX CONCURRENTLY IF NOT EXISTS orders_clobPairId_subaccountId_index ON orders("clobPairId", "subaccountId");
   `);
 }
 
 export async function down(knex: Knex): Promise<void> {
   await knex.raw(`
-    DROP INDEX CONCURRENTLY IF EXISTS "orders_clobpairid_subaccountid_index";
+    DROP INDEX CONCURRENTLY IF EXISTS "orders_clobPairId_subaccountId_index";
   `);
 }